Fix packaging issue for 4.19
This commit will fix packaging issues for UE 4.19

Note:
Not sure if this solves or causes issue for other engine versions
There are changes that might not be necessary, such as swapping out all "UnrealEnginePython.h" header include and the new UEPyMainHeader.h file. Those changes are easy to undo though.