Restore Windows stdio flags after Python initialisation.
Python forcibly sets the file mode for stdin, stdout, and stderr to O_BINARY
for several reasons. This causes wprintf() to output UTF-16 instead of going
through the proper text conversion. As a result, standard output was all
messed up after the module was initialised.

The Python developers do not really consider this a bug because they are not
familiar with wprintf() (https://bugs.python.org/issue16587). Unfortunately
this is what UE4 uses for logging.